+++ src/main/java/com/ercanbeyen/movieapplication/service/impl/MovieServiceImpl.java
@@ -1,6 +1,10 @@
 package com.ercanbeyen.movieapplication.service.impl;
 
-import com.ercanbeyen.movieapplication.constant.OrderBy;
+import com.ercanbeyen.movieapplication.constant.enums.OrderBy;
+import com.ercanbeyen.movieapplication.constant.message.ActionNames;
+import com.ercanbeyen.movieapplication.constant.message.EntityNames;
+import com.ercanbeyen.movieapplication.constant.message.LogMessages;
+import com.ercanbeyen.movieapplication.constant.message.ResponseMessages;
 import com.ercanbeyen.movieapplication.dto.MovieDto;
 import com.ercanbeyen.movieapplication.dto.converter.MovieDtoConverter;
 import com.ercanbeyen.movieapplication.dto.option.filter.MovieFilteringOptions;
@@ -39,6 +43,7 @@ public class MovieServiceImpl implements MovieService {
     @CachePut(value = "movies", key = "#result.id")
     @Override
     public MovieDto createMovie(CreateMovieRequest request) {
+        log.info(String.format(LogMessages.STARTED, "createMovie"));
         Director director = directorService.getDirectorById(request.getDirectorId());
         log.info("Director is found");
 
@@ -52,10 +57,9 @@ public class MovieServiceImpl implements MovieService {
                 .director(director)
                 .actors(new HashSet<>())
                 .build();
-        log.info("Movie is created");
 
         Movie createdMovie = movieRepository.save(newMovie);
-        log.info("Movie is saved");
+        log.info(String.format(LogMessages.SAVED, EntityNames.MOVIE));
 
         return movieDtoConverter.convert(createdMovie);
     }
@@ -63,28 +67,29 @@ public class MovieServiceImpl implements MovieService {
     @CacheEvict(value = "movies", allEntries = true)
     @Override
     public List<MovieDto> getMovies(MovieFilteringOptions filteringOptions, OrderBy orderBy) {
-        log.info("Fetch movies from database");
+        log.info(String.format(LogMessages.STARTED, "getMovies"));
         List<Movie> movies = movieRepository.findAll();
+        log.info(String.format(LogMessages.FETCHED_ALL, EntityNames.MOVIE));
 
         if (filteringOptions.getGenres() != null && !filteringOptions.getGenres().isEmpty()) {
             movies = movies.stream()
                     .filter(movie -> filteringOptions.getGenres().contains(movie.getGenre()))
                     .toList();
-            log.info("Movies are filtered by genres");
+            log.info(String.format(LogMessages.FILTERED, "genre"));
         }
 
         if (!StringUtils.isBlank(filteringOptions.getLanguage())) {
             movies = movies.stream()
                     .filter(movie -> movie.getLanguage().equals(filteringOptions.getLanguage()))
                     .collect(Collectors.toList());
-            log.info("Movies are filtered by language");
+            log.info(String.format(LogMessages.FILTERED, "language"));
         }
 
         if (filteringOptions.getYear() != null) {
             movies = movies.stream()
                     .filter(movie -> movie.getReleaseYear().intValue() == filteringOptions.getYear().intValue())
                     .collect(Collectors.toList());
-            log.info("Movies are filtered by release year");
+            log.info(String.format(LogMessages.FILTERED, "releaseYear"));
         }
 
         if (orderBy != null) {
@@ -101,14 +106,14 @@ public class MovieServiceImpl implements MovieService {
                     })
                     .toList();
 
-            log.info("Movies are sorted by rating");
+            log.info(String.format(LogMessages.SORTED, "rating"));
+        }
 
-            if (filteringOptions.getLimit() != null) {
-                movies = movies.stream()
-                        .limit(filteringOptions.getLimit())
-                        .toList();
-                log.info("Top {} movies are selected", filteringOptions.getLimit());
-            }
+        if (filteringOptions.getLimit() != null) {
+            movies = movies.stream()
+                    .limit(filteringOptions.getLimit())
+                    .toList();
+            log.info(String.format(LogMessages.LIMITED, filteringOptions.getLimit()));
         }
 
         return movies.stream()
@@ -119,18 +124,18 @@ public class MovieServiceImpl implements MovieService {
     @Cacheable(value = "movies", key = "#id", unless = "#result.releaseYear < 2020")
     @Override
     public MovieDto getMovie(Integer id) {
-        log.info("Fetch movie from database");
+        log.info(String.format(LogMessages.STARTED, "getMovie"));
         Movie movieInDb = getMovieById(id);
+        log.info(String.format(LogMessages.FETCHED, EntityNames.MOVIE));
         return movieDtoConverter.convert(movieInDb);
     }
 
     @CacheEvict(value = "movies", allEntries = true)
     @Override
     public MovieDto updateMovie(Integer id, UpdateMovieRequest request) {
-        log.info("Update movie from database");
-
+        log.info(String.format(LogMessages.STARTED, "updateMovie"));
         Movie movieInDb = getMovieById(id);
-        log.info("Movie is found");
+        log.info(String.format(LogMessages.FETCHED, EntityNames.MOVIE));
 
         movieInDb.setTitle(request.getTitle());
         movieInDb.setGenre(request.getGenre());
@@ -138,27 +143,39 @@ public class MovieServiceImpl implements MovieService {
         movieInDb.setReleaseYear(request.getReleaseYear());
         movieInDb.setRating(request.getRating());
         movieInDb.setSummary(request.getSummary());
-        log.info("Fields are set");
+        log.info(LogMessages.FIELDS_SET);
 
-        Movie updatedMovie = movieRepository.save(movieInDb);
-        log.info("Movie is updated");
+        Movie savedMovie = movieRepository.save(movieInDb);
+        log.info(String.format(LogMessages.SAVED, EntityNames.MOVIE));
 
-        return movieDtoConverter.convert(updatedMovie);
+        return movieDtoConverter.convert(savedMovie);
     }
 
     @CacheEvict(value = "movies", key = "#id")
     @Override
     public String deleteMovie(Integer id) {
-        log.info("Delete movie from database");
+        log.info(String.format(LogMessages.STARTED, "deleteMovie"));
+
+        boolean movieExists = movieRepository.existsById(id);
+
+        if (!movieExists) {
+            throw new EntityNotFound(String.format(ResponseMessages.NOT_FOUND, EntityNames.MOVIE, id));
+        }
+
+        log.info(String.format(LogMessages.EXISTS, EntityNames.MOVIE));
         movieRepository.deleteById(id);
-        return "Movie " + id + " is successfully deleted";
+        log.info(String.format(LogMessages.DELETED, EntityNames.MOVIE));
+
+        return String.format(ResponseMessages.SUCCESS, EntityNames.MOVIE, id, ActionNames.DELETED);
     }
 
     @Cacheable(value = "movies")
     @Override
     public List<MovieDto> getLatestMovies() {
-        log.info("Fetch latest movies from database");
+        log.info(String.format(LogMessages.STARTED, "getLatestMovies"));
+
         List<Movie> movies = movieRepository.findAll();
+        log.info(String.format(LogMessages.FETCHED_ALL, EntityNames.MOVIE));
         int year = 2020;
 
         return movies.stream()
@@ -169,11 +186,13 @@ public class MovieServiceImpl implements MovieService {
 
     @Override
     public List<MovieDto> searchMovies(String title) {
+        log.info(String.format(LogMessages.STARTED, "searchMovies"));
         boolean isTitleFilled = StringUtils.isNotBlank(title);
         List<Movie> movies = new ArrayList<>();
 
         if (isTitleFilled) {
             movies = movieRepository.findByTitleStartingWith(title);
+            log.info(String.format(LogMessages.FETCHED_ALL, EntityNames.MOVIE));
         }
 
         return movies.stream()
@@ -183,12 +202,15 @@ public class MovieServiceImpl implements MovieService {
 
     @Override
     public Movie getMovieById(Integer id) {
+        log.info(String.format(LogMessages.STARTED, "getMovieById"));
         return movieRepository.findById(id)
-                .orElseThrow(() -> new EntityNotFound("Movie " + id + " is not found"));
+                .orElseThrow(() -> new EntityNotFound(String.format(ResponseMessages.NOT_FOUND, EntityNames.MOVIE, id)));
     }
 
     @Override
     public CustomPage<MovieDto, Movie> getMovies(Pageable pageable) {
+        log.info(String.format(LogMessages.STARTED, "getMovies"));
+
         Page<Movie> page = movieRepository.findAll(pageable);
         List<MovieDto> movieDtoList = page.getContent().stream()
                 .map(movieDtoConverter::convert)
